+++ frontend/src/app/pages/MainPage/Blocks/ExplorerBlock.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import { styled } from '../../../styles'
+import { styled } from '../../../../styles'
 import sandbox from './img/sandbox.png'
 import arrowLeft from './img/arrow-left.svg'
 
 import sandboxTitle from './img/sandbox-title.svg'
 
 import { BlockWrapper, Subtitle } from './ToolsBlock'
-import { Button } from '../../UIkit'
+import { Button } from '../../../UIkit'
 
 const ExplorerImg = styled('img', {
   width: '100%',
