+++ frontend/src/app/pages/MainPage/Blocks/ToolsBlock.tsx
