+++ frontend/src/app/pages/MainPage/Blocks/WelcomeBlock.tsx
