+++ frontend/src/app/pages/MainPage/MainPage.tsx
@@ -2,7 +2,7 @@ import ToolsBlock from './ToolsBlock'
 import WelcomeBlock from './WelcomeBlock'
 import gradient from './img/Gradient.jpg'
 import { styled } from '../../../styles'
-import ExplorerBlock from './ExplorerBlock'
+import ExplorerBlock from './Blocks/ExplorerBlock'
 
 const GradientWrapper = styled('div', {
   backgroundImage: `url(${gradient})`,
